requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
boto3==1.26.27
botocore==1.29.27
cachetools==5.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cryptography==41.0.3
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
Flask==3.1.0
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.21.0
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.0
jmespath==1.0.1
keyboard==0.13.5
MarkupSafe==3.0.2
MyApplication==0.1.0
numpy
oauth2client==4.1.3
openai==1.55.3
openpyxl==3.1.5
oracledb==1.4.1
packaging==24.2
pandas==2.2.2
pdfkit==1.0.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
pyobjc; sys_platform == 'darwin'
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.13
webvtt-py==0.4.6
Werkzeug==3.1.3
